@@ -1,17 +1,24 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:pfe/services/api_service.dart';
+import 'package:pfe/services/websocket_service.dart';
+import 'package:pfe/models/message.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:pfe/models/chat_group.dart';
+import 'dart:async';
+import 'dart:convert';
 
 class ChatScreen extends StatefulWidget {
-  final String name;
-  final String image;
-  final List<Map<String, dynamic>>? messages;
+  final String userId;
+  final String username;
+  final String? profilePicture;
 
   const ChatScreen({
     super.key,
-    required this.name,
-    required this.image,
-    this.messages,
+    required this.userId,
+    required this.username,
+    this.profilePicture,
   });
 
   @override
@@ -19,178 +26,382 @@ class ChatScreen extends StatefulWidget {
 }
 
 class _ChatScreenState extends State<ChatScreen> {
-  late List<Map<String, dynamic>> messages;
   final TextEditingController _messageController = TextEditingController();
+  final ScrollController _scrollController = ScrollController();
+  final ApiService _apiService = ApiService();
+  final WebSocketService _webSocketService = WebSocketService();
+  List<Message> _messages = [];
+  bool _isLoading = true;
+  String? _currentUserId;
+  String? _currentUsername;
+  StreamSubscription? _chatSubscription;
 
   @override
   void initState() {
     super.initState();
-    // Initialize with provided messages or use default ones
-    messages = widget.messages ?? [
-      {"text": "Hello!", "isSentByUser": false, "date": "Today"},
-      {"text": "How can I help you?", "isSentByUser": false},
-    ];
+    print('[ChatScreen] initState called');
+    print('userId: "${widget.userId}"'); // Debug log
+    print('username: "${widget.username}"'); // Debug log
+    print('profilePicture: "${widget.profilePicture}"'); // Debug log
+    _loadUserData();
+    _initializeChat();
+    
+    // Subscribe to websocket messages
+    _subscribeToWebSocketMessages();
   }
 
-  void _sendMessage() {
-    if (_messageController.text.trim().isNotEmpty) {
-      setState(() {
-        messages.add({"text": _messageController.text, "isSentByUser": true});
-        _messageController.clear();
-      });
+  @override
+  void dispose() {
+    _messageController.dispose();
+    _scrollController.dispose();
+    _chatSubscription?.cancel();
+    super.dispose();
+  }
 
-      // Simulate backend response
-      Future.delayed(Duration(seconds: 2), () {
+  // Subscribe to WebSocket messages
+  void _subscribeToWebSocketMessages() {
+    _chatSubscription = _webSocketService.chatMessageStream.listen((message) {
+      // Check if this message belongs to this chat conversation
+      if (message.chatGroupId == generateChatGroupId(_currentUserId!, widget.userId)) {
         setState(() {
-          messages.add({"text": "Got it!", "isSentByUser": false});
+          // Add the message if it's not already in the list
+          if (!_messages.any((m) => m.id == message.id)) {
+            _messages.add(message);
+            _messages.sort((a, b) => a.timestamp.compareTo(b.timestamp));
+          }
         });
+        
+        // Scroll to bottom after new message is received
+        WidgetsBinding.instance.addPostFrameCallback((_) {
+          _scrollToBottom();
+        });
+      }
+    });
+  }
+
+  Future<void> _loadUserData() async {
+    print('[ChatScreen] _loadUserData called');
+    try {
+      final prefs = await SharedPreferences.getInstance();
+      final userId = prefs.getString('userId');
+      final username = prefs.getString('username');
+      
+      print('Loaded user data:'); // Debug log
+      print('userId: "$userId"'); // Debug log
+      print('username: "$username"'); // Debug log
+      
+      if (mounted) {
+        setState(() {
+          _currentUserId = userId;
+          _currentUsername = username;
+        });
+      }
+    } catch (e) {
+      print('Error loading user data: $e');
+    }
+  }
+
+  Future<void> _initializeChat() async {
+    print('[ChatScreen] _initializeChat called');
+    setState(() {
+      _isLoading = true;
+    });
+    
+    try {
+      await _loadUserData();
+      await _loadMessages();
+      
+      if (_webSocketService.isConnected) {
+        print('WebSocket already connected');
+      } else {
+        print('Connecting to WebSocket...');
+        await _webSocketService.connect();
+      }
+      
+      // Subscribe to the chat topic after connecting
+      final chatGroupId = generateChatGroupId(_currentUserId!, widget.userId);
+      print('Subscribing to chat topic: $chatGroupId');
+      _webSocketService.subscribeToChatGroup(chatGroupId);
+    } catch (e) {
+      print('Error initializing chat: $e');
+    } finally {
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    }
+  }
+
+  Future<void> _loadMessages() async {
+    if (_currentUserId == null) {
+      print('Current user ID is null, cannot load messages');
+      return;
+    }
+    
+    try {
+      final groupId = generateChatGroupId(_currentUserId!, widget.userId);
+      print('[ChatScreen] Loading messages for chat group: $groupId');
+      
+      final messages = await _apiService.getIndividualMessages(
+        _currentUserId!,
+        widget.userId,
+      );
+      
+      if (mounted) {
+        setState(() {
+          _messages = messages;
+          _isLoading = false;
+        });
+        
+        WidgetsBinding.instance.addPostFrameCallback((_) {
+          _scrollToBottom();
+        });
+      }
+    } catch (e) {
+      print('[ChatScreen] Error loading messages: $e');
+      if (mounted) {
+        setState(() {
+          _isLoading = false;
+        });
+      }
+    }
+  }
+
+  String generateChatGroupId(String userId1, String userId2) {
+    List<String> ids = [userId1, userId2];
+    ids.sort(); // Sort to ensure the same ID regardless of order
+    return 'individual_${ids[0]}_${ids[1]}';
+  }
+
+  void _scrollToBottom() {
+    if (_scrollController.hasClients) {
+      _scrollController.animateTo(
+        _scrollController.position.maxScrollExtent,
+        duration: const Duration(milliseconds: 300),
+        curve: Curves.easeOut,
+      );
+    }
+  }
+
+  Future<void> _sendMessage() async {
+    print('[ChatScreen] _sendMessage called');
+    if (_messageController.text.trim().isEmpty) return;
+    
+    final String content = _messageController.text.trim();
+    
+    // Create a temporary message for the UI
+    final tempMessage = Message(
+      id: 'temp_${DateTime.now().millisecondsSinceEpoch}',
+      senderId: _currentUserId!,
+      receiverId: widget.userId,
+      chatGroupId: generateChatGroupId(_currentUserId!, widget.userId),
+      content: content,
+      timestamp: DateTime.now(),
+      senderName: _currentUsername ?? 'Me',
+    );
+    
+    setState(() {
+      _messages.add(tempMessage);
+      _messageController.clear();
+    });
+    
+    // Scroll to bottom after message is added
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      _scrollToBottom();
+    });
+
+    try {
+      // Send message via REST API
+      final message = await _apiService.sendIndividualMessage(
+        _currentUserId!,
+        widget.userId,
+        content,
+      );
+
+      // Also send via WebSocket for real-time delivery
+      await _webSocketService.sendChatMessage(widget.userId, content);
+
+      // Replace the temporary message with the real one
+      setState(() {
+        final index = _messages.indexWhere((m) => 
+          m.id.startsWith('temp_') && 
+          m.content == content && 
+          m.senderId == _currentUserId);
+          
+        if (index != -1) {
+          _messages[index] = message;
+        }
       });
+    } catch (e) {
+      print('Error sending message: $e');
+      
+      // Remove the temporary message
+      setState(() {
+        _messages.removeWhere((m) => 
+          m.id.startsWith('temp_') && 
+          m.content == content && 
+          m.senderId == _currentUserId);
+      });
+      
+      // Show error message
+      ScaffoldMessenger.of(context).showSnackBar(
+        SnackBar(
+          content: Text('Failed to send message: $e'),
+          backgroundColor: Colors.red,
+        ),
+      );
     }
   }
 
   @override
   Widget build(BuildContext context) {
+    print('[ChatScreen] build called');
     return Scaffold(
-      backgroundColor: Color(0xFFF0F7F5),
-      
-      appBar: PreferredSize(
-      preferredSize: Size.fromHeight(100), // Keep the height as required
-      child: Container(
-        decoration: BoxDecoration(
-          color: Color(0xFFF0F7F5), // Match background color
-          border: Border(
-            bottom: BorderSide(
-              color: Colors.black.withOpacity(0.1), // 10% opacity black line
-              width: 1, // Thin line
-            ),
-          ),
+      backgroundColor: const Color(0xFFF0F7F5),
+      appBar: AppBar(
+        backgroundColor: Colors.white,
+        elevation: 0,
+        leadingWidth: 56,
+        leading: IconButton(
+          icon: const Icon(FontAwesomeIcons.arrowLeft, color: Color(0xC5000000)),
+          onPressed: () => Navigator.of(context).pop(),
         ),
-        child: AppBar(
-          backgroundColor: Colors.transparent, // Transparent AppBar to match container
-          elevation: 0, // Remove shadow to keep only the thin line
-          toolbarHeight: 100,
-          leading: IconButton(
-            icon: Icon(FontAwesomeIcons.arrowLeft, color: Color(0xC5000000)),
-            onPressed: () => Navigator.pop(context),
-          ),
-          title: Row(
-            children: [
-              CircleAvatar(
-                radius: 22, // Adjust size if needed
-                backgroundImage: AssetImage(widget.image),
+        title: Row(
+          children: [
+            CircleAvatar(
+              radius: 16,
+              backgroundColor: Colors.grey[300],
+              backgroundImage: widget.profilePicture != null && widget.profilePicture!.isNotEmpty
+                  ? MemoryImage(base64Decode(widget.profilePicture!.split(',').last)) as ImageProvider
+                  : null,
+              child: widget.profilePicture == null || widget.profilePicture!.isEmpty
+                  ? Text(
+                      widget.username[0].toUpperCase(),
+                      style: const TextStyle(
+                        color: Colors.white,
+                        fontWeight: FontWeight.bold,
+                      ),
+                    )
+                  : null,
+            ),
+            const SizedBox(width: 8),
+            Text(
+              widget.username,
+              style: GoogleFonts.poppins(
+                fontSize: 18,
+                fontWeight: FontWeight.w500,
+                color: const Color(0xC5000000),
               ),
-              SizedBox(width: 10), // Space between profile picture and name
-              Text(
-                widget.name,
-                style: GoogleFonts.poppins(
-                  fontSize: 18, 
-                  fontWeight: FontWeight.w500, 
-                  color: Color(0xC5000000),
-                ),
-              ),
-            ],
-          ),
+            ),
+          ],
         ),
       ),
-    ),
       body: Column(
         children: [
           Expanded(
-            child: ListView.builder(
-              padding: EdgeInsets.all(10),
-              itemCount: messages.length,
-              itemBuilder: (context, index) {
-                final message = messages[index];
-                bool isSentByUser = message["isSentByUser"];
-                bool isDateMessage = message.containsKey("date");
-
-                return isDateMessage
-                    ? Center(
-                        child: Padding(
-                          padding: const EdgeInsets.symmetric(vertical: 8),
-                          child: Text(
-                            message["date"],
-                            style: GoogleFonts.poppins(color: Color(0x45000000), fontSize: 14),
+            child: _isLoading
+                ? const Center(child: CircularProgressIndicator())
+                : _messages.isEmpty
+                    ? const Center(
+                        child: Text(
+                          'No messages yet',
+                          style: TextStyle(
+                            color: Colors.black54,
+                            fontSize: 14,
                           ),
                         ),
                       )
-                    :Align(
-                      alignment: isSentByUser ? Alignment.centerRight : Alignment.centerLeft,
-                      child: Container(
-                        //margin: EdgeInsets.symmetric(vertical: 5, horizontal: 10),
-                        padding: EdgeInsets.all(12),
-                        decoration: BoxDecoration(
-                          color: isSentByUser ? Color(0xFFD0ECE8) : Color(0xFFE4E4E4).withOpacity(0.83),
-                          borderRadius: isSentByUser
-                              ? BorderRadius.only(
-                                  topLeft: Radius.circular(20),
-                                  topRight: Radius.circular(20),
-                                  bottomLeft: Radius.circular(20),
-                                  bottomRight: Radius.circular(0), // Less rounded for a bubble effect
-                                )
-                              : BorderRadius.only(
-                                  topLeft: Radius.circular(20),
-                                  topRight: Radius.circular(20),
-                                  bottomRight: Radius.circular(20),
-                                  bottomLeft: Radius.circular(0), // Less rounded for a bubble effect
+                    : ListView.builder(
+                        padding: const EdgeInsets.all(16),
+                        itemCount: _messages.length,
+                        itemBuilder: (context, index) {
+                          final message = _messages[index];
+                          final isMe = message.senderId == _currentUserId;
+
+                          return Align(
+                            alignment: isMe ? Alignment.centerRight : Alignment.centerLeft,
+                            child: Container(
+                              padding: const EdgeInsets.all(12),
+                              margin: const EdgeInsets.only(bottom: 8),
+                              decoration: BoxDecoration(
+                                color: isMe ? const Color(0xFFD0ECE8) : const Color(0xFFF5F5F5),
+                                borderRadius: BorderRadius.only(
+                                  topLeft: const Radius.circular(16),
+                                  topRight: const Radius.circular(16),
+                                  bottomLeft: Radius.circular(isMe ? 16 : 0),
+                                  bottomRight: Radius.circular(isMe ? 0 : 16),
                                 ),
-                        ),
-                        child: Text(
-                          message["text"],
-                          style: GoogleFonts.poppins(fontSize: 14, color: Color(0xC5000000)), // Adjust text style if needed
+                              ),
+                              constraints: BoxConstraints(
+                                maxWidth: MediaQuery.of(context).size.width * 0.75,
+                              ),
+                              child: Text(
+                                message.content,
+                                style: const TextStyle(
+                                  fontSize: 14,
+                                  color: Colors.black87,
+                                ),
+                              ),
+                            ),
+                          );
+                        },
+                      ),
+          ),
+          Container(
+            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+            decoration: BoxDecoration(
+              color: Colors.white,
+              border: Border(
+                top: BorderSide(
+                  color: Colors.grey.withOpacity(0.2),
+                ),
+              ),
+            ),
+            child: SafeArea(
+              child: Row(
+                children: [
+                  Expanded(
+                    child: Container(
+                      padding: const EdgeInsets.symmetric(horizontal: 16),
+                      decoration: BoxDecoration(
+                        color: const Color(0xFFD0ECE8).withOpacity(0.3),
+                        borderRadius: BorderRadius.circular(24),
+                      ),
+                      child: TextField(
+                        controller: _messageController,
+                        decoration: const InputDecoration(
+                          hintText: 'Send Message',
+                          hintStyle: TextStyle(color: Colors.black54),
+                          border: InputBorder.none,
+                          contentPadding: EdgeInsets.symmetric(vertical: 10),
                         ),
                       ),
-                    );
-
-              },
+                    ),
+                  ),
+                  const SizedBox(width: 8),
+                  Material(
+                    color: const Color(0xFF6BBFB5),
+                    borderRadius: BorderRadius.circular(24),
+                    child: InkWell(
+                      onTap: _sendMessage,
+                      borderRadius: BorderRadius.circular(24),
+                      child: const Padding(
+                        padding: EdgeInsets.all(10),
+                        child: Icon(
+                          Icons.send,
+                          color: Colors.white,
+                          size: 20,
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
             ),
           ),
-          _buildMessageInput(),
         ],
       ),
     );
   }
-
-  Widget _buildMessageInput() {
-  return Container(
-    padding: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
-    decoration: BoxDecoration(
-      color: Color(0xFFF0F7F5),
-      boxShadow: [BoxShadow(color: Colors.black12, blurRadius: 4)],
-    ),
-    child: Row(
-      children: [
-        SizedBox(
-          width: 286, // Set width to 286
-          height: 54, // Set height to 54
-          child: TextField(
-            controller: _messageController,
-            decoration: InputDecoration(
-              hintText: "Send Message",
-              filled: true,
-              fillColor: Color(0xFFB1DDD4).withOpacity(0.5),
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(28),
-                borderSide: BorderSide.none,
-              ),
-              contentPadding: EdgeInsets.symmetric(horizontal: 30, vertical: 16),
-            ),
-          ),
-        ),
-
-        SizedBox(width: 12),
-
-        GestureDetector(
-          onTap: _sendMessage,
-          child: CircleAvatar(
-            radius: 25, // Increase radius if needed
-            backgroundColor: Color(0xFF6BBFB5),
-            child: Icon(Icons.send, color: Colors.white, weight:25)
-          ),
-        ),
-      ],
-    ),
-  );
-}
-
 }
