@@ -3,9 +3,7 @@ import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:pfe/controllers/signup_controller.dart';
 import 'package:pfe/screens/authentication/otp_verification_screen.dart';
-import 'package:pfe/services/auth_service.dart';
-
-
+import 'package:pfe/services/api_service.dart';
 
 class SignUpScreen extends StatefulWidget {
   const SignUpScreen({super.key});
@@ -17,76 +15,87 @@ class SignUpScreen extends StatefulWidget {
 class _SignUpScreenState extends State<SignUpScreen> {
   final SignUpControllers controllers = SignUpControllers();
   final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-  final _authService = AuthService();
+  final _apiService = ApiService();
   bool _isLoading = false;
-
-
   bool _isPasswordVisible = false;
   bool _isConfirmPasswordVisible = false;
 
   void _handleSignUp() async {
     if (_formKey.currentState!.validate()) {
+      if (controllers.selectedRole.value == null) {
+        ScaffoldMessenger.of(context).showSnackBar(
+          const SnackBar(
+            content: Text('Please select a role'),
+            backgroundColor: Colors.red,
+          ),
+        );
+        return;
+      }
+
       setState(() => _isLoading = true);
-      bool success = await _authService.signUp(
-        name: controllers.nameController.text.trim(),
-        email: controllers.emailController.text.trim(),
-        password: controllers.passwordController.text,
-      );
-    
 
-      setState(() => _isLoading = false);
-
-      if (success) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(content: Text('Signup successful!')),
-
+      try {
+        final message = await _apiService.signup(
+          controllers.emailController.text.trim(),
+          controllers.nameController.text.trim(),
+          controllers.passwordController.text,
+          controllers.selectedRole.value!,
         );
+
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              content: Text(message),
+              backgroundColor: Colors.green,
+            ),
+          );
+          
           Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                              builder: (context) => OtpVerificationScreen(email: controllers.emailController.text),
-                            ),
-                          );
-      } else {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(content: Text('Signup failed!')),
-        );
+            context,
+            MaterialPageRoute(
+              builder: (context) => OtpVerificationScreen(
+                email: controllers.emailController.text.trim(),
+              ),
+            ),
+          );
+        }
+      } catch (e) {
+        if (mounted) {
+          ScaffoldMessenger.of(context).showSnackBar(
+            SnackBar(
+              content: Text(e.toString().replaceAll('Exception: ', '')),
+              backgroundColor: Colors.red,
+            ),
+          );
+        }
+      } finally {
+        if (mounted) {
+          setState(() => _isLoading = false);
+        }
       }
     }
   }
 
   @override
   void dispose() {
-    controllers.nameController.dispose();
-    controllers.emailController.dispose();
-    controllers.passwordController.dispose();
-    controllers.confirmPasswordController.dispose();
+    controllers.dispose();
     super.dispose();
   }
-
-
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-
       backgroundColor: const Color(0xFFF0F7F5),
-
       body: SafeArea(
         child: SingleChildScrollView(
           child: Padding(
             padding: const EdgeInsets.all(24.0),
             child: Form(
               key: _formKey,
-              
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.stretch,
                 children: [
-
-
                   const SizedBox(height: 100),
-
-
                   Text(
                     'Welcome Onboard',
                     style: GoogleFonts.poppins(
@@ -96,11 +105,7 @@ class _SignUpScreenState extends State<SignUpScreen> {
                     ),
                     textAlign: TextAlign.center,
                   ),
-
-
                   const SizedBox(height: 10),
-
-
                   Text(
                     "Let's help you perform better everyday",
                     style: GoogleFonts.pontanoSans(
@@ -109,164 +114,180 @@ class _SignUpScreenState extends State<SignUpScreen> {
                     ),
                     textAlign: TextAlign.center,
                   ),
-
-
                   const SizedBox(height: 39),
-                  
-
-                      TextFormField(
-                        controller: controllers.nameController,
-                        decoration: InputDecoration(
-                          hintText: 'Enter your full name',
-                          hintStyle: GoogleFonts.pontanoSans(
-                            fontSize: 14,
-                            fontWeight: FontWeight.w600,
-                          ),
-                          filled: true,
-                          fillColor: Color(0xFFF0F0F0),
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(12),
-                            borderSide: BorderSide.none,
-                          ),
-                          contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
-                        ),
+                  TextFormField(
+                    controller: controllers.nameController,
+                    decoration: InputDecoration(
+                      hintText: 'Enter your full name',
+                      hintStyle: GoogleFonts.pontanoSans(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
                       ),
-
-
-                      const SizedBox(height: 20),
-
-
-                      TextFormField(
-                        controller: controllers.emailController,
-                        decoration: InputDecoration(
-                          hintText: 'Enter your email',
-                          hintStyle: GoogleFonts.pontanoSans(
-                            fontSize: 14,
-                            fontWeight: FontWeight.w600,
-                          ),
-                          filled: true,
-                          fillColor: Color(0xFFF0F0F0),
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(12),
-                            borderSide: BorderSide.none,
-                          ),
-                          contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
-                        ),
-                        validator: controllers.validateEmail,
+                      filled: true,
+                      fillColor: Color(0xFFF0F0F0),
+                      border: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(12),
+                        borderSide: BorderSide.none,
                       ),
-
-
-                      const SizedBox(height: 20),
-
-
-                      TextFormField(
-                        controller: controllers.passwordController,
-                        obscureText: !_isPasswordVisible,  // Toggle visibility
+                      contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
+                    ),
+                  ),
+                  const SizedBox(height: 20),
+                  TextFormField(
+                    controller: controllers.emailController,
+                    decoration: InputDecoration(
+                      hintText: 'Enter your email',
+                      hintStyle: GoogleFonts.pontanoSans(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                      ),
+                      filled: true,
+                      fillColor: Color(0xFFF0F0F0),
+                      border: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(12),
+                        borderSide: BorderSide.none,
+                      ),
+                      contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
+                    ),
+                    validator: controllers.validateEmail,
+                  ),
+                  const SizedBox(height: 20),
+                  TextFormField(
+                    controller: controllers.passwordController,
+                    obscureText: !_isPasswordVisible,
+                    decoration: InputDecoration(
+                      hintText: 'Enter password',
+                      hintStyle: GoogleFonts.pontanoSans(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                      ),
+                      filled: true,
+                      fillColor: const Color(0xFFF0F0F0),
+                      border: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(12),
+                        borderSide: BorderSide.none,
+                      ),
+                      contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
+                      suffixIcon: IconButton(
+                        icon: FaIcon(
+                          _isPasswordVisible 
+                            ? FontAwesomeIcons.eye
+                            : FontAwesomeIcons.eyeSlash,
+                          color: Colors.grey,
+                        ),
+                        onPressed: () {
+                          setState(() {
+                            _isPasswordVisible = !_isPasswordVisible;
+                          });
+                        },
+                      ),
+                    ),
+                    validator: controllers.validatePassword,
+                  ),
+                  const SizedBox(height: 20),
+                  TextFormField(
+                    controller: controllers.confirmPasswordController,
+                    obscureText: !_isConfirmPasswordVisible,
+                    decoration: InputDecoration(
+                      hintText: 'Confirm password',
+                      hintStyle: GoogleFonts.pontanoSans(
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600,
+                      ),
+                      filled: true,
+                      fillColor: const Color(0xFFF0F0F0),
+                      border: OutlineInputBorder(
+                        borderRadius: BorderRadius.circular(12),
+                        borderSide: BorderSide.none,
+                      ),
+                      contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
+                      suffixIcon: IconButton(
+                        icon: FaIcon(
+                          _isConfirmPasswordVisible
+                            ? FontAwesomeIcons.eye
+                            : FontAwesomeIcons.eyeSlash,
+                          color: Colors.grey,
+                        ),
+                        onPressed: () {
+                          setState(() {
+                            _isConfirmPasswordVisible = !_isConfirmPasswordVisible;
+                          });
+                        },
+                      ),
+                    ),
+                    validator: controllers.validateConfirmPassword,
+                  ),
+                  const SizedBox(height: 20),
+                  Container(
+                    padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 5),
+                    decoration: BoxDecoration(
+                      color: Color(0xFFF0F0F0),
+                      borderRadius: BorderRadius.circular(12),
+                    ),
+                    child: DropdownButtonHideUnderline(
+                      child: DropdownButtonFormField<String>(
+                        value: controllers.selectedRole.value,
                         decoration: InputDecoration(
-                          hintText: 'Enter password',
-                          hintStyle: GoogleFonts.pontanoSans(
+                          border: InputBorder.none,
+                          contentPadding: EdgeInsets.zero,
+                          errorStyle: TextStyle(color: Colors.red),
+                        ),
+                        hint: Text(
+                          'Select Role',
+                          style: GoogleFonts.pontanoSans(
                             fontSize: 14,
                             fontWeight: FontWeight.w600,
                           ),
-                          filled: true,
-                          fillColor: const Color(0xFFF0F0F0),
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(12),
-                            borderSide: BorderSide.none,
-                          ),
-                          contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
-                          suffixIcon: IconButton(
-                            icon: FaIcon(
-                              _isPasswordVisible 
-                                ? FontAwesomeIcons.eye // Icon when visible
-                                : FontAwesomeIcons.eyeSlash, // Icon when hidden
-                              color: Colors.grey,
+                        ),
+                        items: ['TECHNICIAN', 'SUPERVISOR', 'MANAGER'].map((String role) {
+                          return DropdownMenuItem<String>(
+                            value: role,
+                            child: Text(
+                              role[0] + role.substring(1).toLowerCase(),
+                              style: GoogleFonts.pontanoSans(
+                                fontSize: 14,
+                                fontWeight: FontWeight.w600,
+                              ),
                             ),
-                            onPressed: () {
-                              setState(() {
-                                _isPasswordVisible = !_isPasswordVisible; // Toggle visibility
-                              });
-                            },
-                          ),
-                        ),
-                        validator: controllers.validatePassword,
+                          );
+                        }).toList(),
+                        onChanged: (String? newValue) {
+                          if (newValue != null) {
+                            setState(() {
+                              controllers.selectedRole.value = newValue;
+                            });
+                          }
+                        },
+                        validator: controllers.validateRole,
+                        isExpanded: true,
                       ),
-
-
-
-                      const SizedBox(height: 20),
-
-
-
-                      TextFormField(
-                        controller: controllers.confirmPasswordController,
-                        obscureText: !_isConfirmPasswordVisible,  // Toggle visibility
-                        decoration: InputDecoration(
-                          hintText: 'Confirm password',
-                          hintStyle: GoogleFonts.pontanoSans(
-                            fontSize: 14,
-                            fontWeight: FontWeight.w600,
-                          ),
-                          filled: true,
-                          fillColor: const Color(0xFFF0F0F0),
-                          border: OutlineInputBorder(
-                            borderRadius: BorderRadius.circular(12),
-                            borderSide: BorderSide.none,
-                          ),
-                          contentPadding: const EdgeInsets.symmetric(horizontal: 20, vertical: 15),
-                          suffixIcon: IconButton(
-                            icon: FaIcon(
-                              _isConfirmPasswordVisible
-                                ? FontAwesomeIcons.eye  // Icon when visible
-                                : FontAwesomeIcons.eyeSlash,     // Icon when hidden
-                              color: Colors.grey,
-                            ),
-                            onPressed: () {
-                              setState(() {
-                                _isConfirmPasswordVisible = !_isConfirmPasswordVisible; // Toggle visibility
-                              });
-                            },
-                          ),
-                        ),
-                        validator: controllers.validateConfirmPassword,
-                      ),
-                                      
-
-
+                    ),
+                  ),
                   const SizedBox(height: 45),
-
-
-
                   SizedBox(
                     width: double.infinity,
                     height: 55,
                     child: 
                     _isLoading
-                  ? const CircularProgressIndicator()
-                  :ElevatedButton(
-                      onPressed: () {
-                        _handleSignUp();
-                      },
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: const Color(0xFF6BBFB5),
-                        padding: const EdgeInsets.symmetric(vertical: 15),
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(10.0),
+                    ? const CircularProgressIndicator()
+                    : ElevatedButton(
+                        onPressed: () {
+                          _handleSignUp();
+                        },
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: const Color(0xFF6BBFB5),
+                          padding: const EdgeInsets.symmetric(vertical: 15),
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(10.0),
+                          ),
+                        ),
+                        child: const Text(
+                          'Sign up',
+                          style: TextStyle(color: Colors.white, fontSize: 18),
                         ),
                       ),
-                      child: const Text(
-                        'Sign up',
-                        style: TextStyle(color: Colors.white, fontSize: 18),
-                      ),
-                    ),
                   ),
-
-
-
                   const SizedBox(height: 20),
-
-
-                  
                   Row(
                     mainAxisAlignment: MainAxisAlignment.center,
                     children: [
