@@ -8,9 +8,9 @@ import 'manager_notification_view.dart';
 import 'received_notifications_view.dart';
 
 enum UserRole {
-  manager,
-  supervisor,
-  Technician
+  MANAGER,
+  SUPERVISOR,
+  TECHNICIAN
 }
 
 class RoleBasedNotificationScreen extends StatefulWidget {
@@ -27,9 +27,26 @@ class RoleBasedNotificationScreen extends StatefulWidget {
 
 class _RoleBasedNotificationScreenState extends State<RoleBasedNotificationScreen> {
   @override
+  void initState() {
+    super.initState();
+    print('RoleBasedNotificationScreen initialized with role: ${widget.userRole}'); // Debug print
+  }
+
+  @override
   Widget build(BuildContext context) {
+    print('Building RoleBasedNotificationScreen with role: ${widget.userRole}'); // Debug print
     return ChangeNotifierProvider(
-      create: (_) => NotificationProvider(),
+      create: (_) {
+        final provider = NotificationProvider();
+        // Force set the role immediately
+        provider.setRole(widget.userRole.toString().split('.').last);
+        // Force technicians to received view
+        if (widget.userRole == UserRole.TECHNICIAN) {
+          provider.isReceived = true;
+        }
+        print('NotificationProvider created with role: ${provider.userRole}'); // Debug print
+        return provider;
+      },
       child: Scaffold(
         backgroundColor: Color(0xFFF0F7F5),
         appBar: AppBar(
@@ -57,29 +74,23 @@ class _RoleBasedNotificationScreenState extends State<RoleBasedNotificationScree
         ),
         body: Column(
           children: [
-            _buildToggleButtons(),
+            // Only show toggle buttons for managers and supervisors
+            if (widget.userRole != UserRole.TECHNICIAN) _buildToggleButtons(),
             Expanded(
               child: Consumer<NotificationProvider>(
                 builder: (context, provider, child) {
-                  if (provider.isReceived) {
+                  // Technicians can only see received notifications
+                  if (widget.userRole == UserRole.TECHNICIAN || provider.isReceived) {
                     return ReceivedNotificationsView();
                   } else {
                     // Show different send views based on user role
                     switch (widget.userRole) {
-                      case UserRole.manager:
+                      case UserRole.MANAGER:
                         return ManagerNotificationView();
-                      case UserRole.supervisor:
+                      case UserRole.SUPERVISOR:
                         return SendNotificationView();
-                      case UserRole.Technician:
-                        return Center(
-                          child: Text(
-                            'You do not have permission to send notifications',
-                            style: GoogleFonts.poppins(
-                              fontSize: 16,
-                              color: Color(0xC5000000),
-                            ),
-                          ),
-                        );
+                      case UserRole.TECHNICIAN:
+                        return ReceivedNotificationsView(); // Fallback to received view
                     }
                   }
                 },
@@ -94,6 +105,36 @@ class _RoleBasedNotificationScreenState extends State<RoleBasedNotificationScree
   Widget _buildToggleButtons() {
     return Consumer<NotificationProvider>(
       builder: (context, provider, child) {
+        // If user is a technician, only show the Received button
+        if (widget.userRole == UserRole.TECHNICIAN) {
+          return Container(
+            margin: EdgeInsets.all(16),
+            decoration: BoxDecoration(
+              color: Color(0xFFE8F3F1),
+              borderRadius: BorderRadius.circular(25),
+            ),
+            child: SizedBox(
+              width: 280,
+              child: Container(
+                padding: EdgeInsets.symmetric(vertical: 12),
+                decoration: BoxDecoration(
+                  color: Color(0xFF6BBFB5),
+                  borderRadius: BorderRadius.circular(25),
+                ),
+                child: Text(
+                  'Received',
+                  textAlign: TextAlign.center,
+                  style: GoogleFonts.poppins(
+                    color: Colors.white,
+                    fontWeight: FontWeight.w500,
+                  ),
+                ),
+              ),
+            ),
+          );
+        }
+
+        // For managers and supervisors, show both toggles
         return Container(
           margin: EdgeInsets.all(16),
           decoration: BoxDecoration(
